Move ConfigConstants to MicroserviceConstants
