--- src/Monq.Core.BasicDotNetMicroservice/Configuration/ConfigConstants.cs
@@ -1,27 +1,0 @@
-namespace Monq.Core.BasicDotNetMicroservice.Configuration;
-
-/// <summary>
-/// Configuration constants.
-/// </summary>
-public static class ConfigConstants
-{
-    /// <summary>
-    /// Redis section.
-    /// </summary>
-    public const string Redis = "Redis";
-
-    /// <summary>
-    /// RabbitMQ section.
-    /// </summary>
-    public const string RabbitMq = "RabbitMQ";
-
-    /// <summary>
-    /// ClickHouse connection string.
-    /// </summary>
-    public const string ClickHouseConnectionString = "ClickHouseHttp:DefaultConnection:ConnectionString";
-
-    /// <summary>
-    /// PostgreSQL connection string.
-    /// </summary>
-    public const string PostgresConnectionString = "PostgreSQL:DefaultConnection:ConnectionString";
-}
